requirements.txt
django==3.1.3
django-ninja==0.20.0
psycopg2-binary==2.9.3
loguru==0.5.3
